requirements.txt
asgiref==3.5.2
attrs==22.1.0
dj-database-url==1.0.0
Django==3.2.16
django-cors-headers==3.13.0
django-currentuser==0.5.3
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.23.1
inflection==0.5.1
jsonschema==4.14.0
Pillow==9.2.0
psycopg2==2.9.3
PyJWT==2.4.0
pyrsistent==0.18.1
pytz==2022.2.1
PyYAML==6.0
sqlparse==0.4.2
tzdata==2022.2
uritemplate==4.1.1
whitenoise==6.2.0
